fix: use opaque type
